Move success message inside if body
With the previous code, regardless of the result of the if statement a
success message will be printed. This means that the user could get both
a failure message and a success message, meaning they got a success
message even when they were not granted seats.

In order to fix this, the success message should be printed within the
body of the else. This way, it will only be printed if there are
sufficient seats.